TileMap rendering up 'n running
commit b8b1f8a92b1e84db79a43fe7ff45df9844b81657
Author: Ilari Tommiska <ilari97@hotmail.com>
Date:   Fri Nov 22 09:14:24 2019 +0200

    Simple tests for TileMap

commit 1045b0262ab5ea7b20352e6e10da36ab9f6674e9
Author: Ilari Tommiska <ilari97@hotmail.com>
Date:   Thu Nov 21 14:37:04 2019 +0200

    Aggregate JaCoCo report

commit ac4c31984848d888d75be7b18a8faf261ed859e9
Author: Ilari Tommiska <ilari97@hotmail.com>
Date:   Thu Nov 21 13:57:39 2019 +0200

    Fixed tests to expect AssertionErrors from NonNull

commit 546a5b6376d2b1f6d38b5946c2625cbed0fb8e0d
Author: Lassi Kuisma <lassi.kuisma1@gmail.com>
Date:   Wed Nov 20 13:59:08 2019 +0200

    Implemented a basic spawner entity. Added unit test for stalker enemy.

commit 556875ea38296f63be6458331a2884f1052d78dc
Author: Ilari Tommiska <ilari97@hotmail.com>
Date:   Thu Nov 21 11:53:03 2019 +0200

    Camera now follows player

commit 5f67b20f8152f55cd4e5246f822c4f922dba845c
Author: Ilari Tommiska <ilari97@hotmail.com>
Date:   Thu Nov 21 10:08:18 2019 +0200

    Chunky level storage first draft

commit a1cd7ab616d5069e8aa365342088b898b273d12c
Author: Ilari Tommiska <ilari97@hotmail.com>
Date:   Thu Nov 21 07:34:36 2019 +0200

    TileMap rendering now uses streams to iterate regions

commit 39833ea8b83e28aa7fe4c3a6fe03b0d7fa3518c6
Author: Ilari Tommiska <ilari97@hotmail.com>
Date:   Wed Nov 20 14:58:16 2019 +0200

    Reworked sprite asset handling, got TileMap renderer first draft working

+++ engine/src/main/java/fi/jakojaannos/roguelite/engine/view/Camera.java
@@ -1,31 +1,31 @@
 package fi.jakojaannos.roguelite.engine.view;
 
-import org.joml.Vector2f;
+import org.joml.Vector2d;
 
 public class Camera {
-    private final Vector2f position;
+    private final Vector2d position;
 
-    public final float getX() {
+    public final double getX() {
         return this.position.x;
     }
 
-    public final float getY() {
+    public final double getY() {
         return this.position.y;
     }
 
-    public final void setX(float x) {
+    public final void setX(double x) {
         setPosition(x, getY());
     }
 
-    public final void setY(float y) {
+    public final void setY(double y) {
         setPosition(getX(), y);
     }
 
-    public void setPosition(float x, float y) {
+    public void setPosition(double x, double y) {
         this.position.set(x, y);
     }
 
-    public Camera(Vector2f position) {
-        this.position = new Vector2f(position);
+    public Camera(Vector2d position) {
+        this.position = new Vector2d(position);
     }
 }
